% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_quantity.R
\name{animal_quantity}
\alias{animal_quantity}
\title{Calculates the quantity of marine animals}
\usage{
animal_quantity(whales, turtles)
}
\arguments{
\item{whales}{number of whales}

\item{turtles}{number of turtles}
}
\description{
Calculates the quantity of marine animals
}
\examples{
animal_quantity(whale = 30, turtle = 22)
}
