% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meals.R
\name{meals}
\alias{meals}
\title{Calculates what you like to eat at specific times}
\usage{
meals(food, time)
}
\arguments{
\item{time}{}
}
\description{
Calculates what you like to eat at specific times
}
\examples{
meal(food ="pasta", time = 3)
}
